% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-functions.R
\name{createExperimentsFragmentOptimisation}
\alias{createExperimentsFragmentOptimisation}
\title{Create fragment optimisation experiment}
\usage{
createExperimentsFragmentOptimisation(ms1, ..., groupBy = c("AgcTarget",
  "replication"), nMs2perMs1 = 10, duration = 0.5, replications = 2,
  randomise = TRUE, massLabeling = TRUE)
}
\arguments{
\item{ms1}{\code{data.frame}, MS1 settings.}

\item{...}{further named arguments with \code{data.frame}s containing the MS2
settings.}

\item{nMs2perMs1}{\code{integer},}

\item{duration}{\code{double}, how long should the scan be?}

\item{replications}{\code{integer}, number of replications.}

\item{randomise}{\code{logical}, should the MS2 scan settings randomised?}

\item{massLabeling}{\code{logical},
how many MS2 scans should be run after a MS1 scan?}

\item{groupyBy}{\code{character}, group experiments by columns in the MS2
\code{data.frame}s. The columns have to be present in all \code{data.frame}s. Each
group will be written to its own XML file.}
}
\value{
\code{list}, able to be written via \code{\link[xml2:as_xml_document]{xml2::as_xml_document()}}
}
\description{
Create fragment optimisation experiment
}
