% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-functions.R
\name{defaultMs1Settings}
\alias{defaultMs1Settings}
\alias{defaultMs2Settings}
\alias{defaultProteins}
\title{Settings for MS1/2 parameters.}
\usage{
defaultMs1Settings(...)

defaultMs2Settings(...)

defaultProteins(protein = c("GST", "myoglobin", "h2a", "h2b", "h3_1", "h3_3",
  "h4", "c345c_c3", "h33tail"))
}
\arguments{
\item{\ldots}{named arguments that should be overwritten.}
}
\value{
A named \code{list} of settings.
}
\description{
This functions create the default settings for \code{\link[=writeMethodXmls]{writeMethodXmls()}}.
}
\examples{
library("topdown")

# all default MS1 settings
defaultMs1Settings()

# overwrite FirstMass and set AgcTarget
defaultMs1Settings(FirstMass=100, AgcTarget=c(1e4, 1e5))

# all default MS2 settings
defaultMs2Settings()

# overwrite AgcTarget
defaultMs2Settings(AgcTarget=c(1e4, 1e5))

# default mass/z values for H2A
defaultProteins("h2a")
}
