% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings-functions.R
\name{defaultMs1Settings}
\alias{defaultMs1Settings}
\title{Settings for MS1 parameters.}
\usage{
defaultMs1Settings(...)
}
\arguments{
\item{\ldots}{named arguments that should be overwritten}
}
\value{
named \code{list} of settings
}
\description{
Settings for MS1 parameters.
}
\examples{
library("topdown")

# all default settings
defaultMs1Settings()

# overwrite FirstMass and set AgcTarget
defaultMs1Settings(FirstMass=100, AgcTarget=c(1e4, 1e5))
}
