% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings-functions.R
\name{defaultMs2Settings}
\alias{defaultMs2Settings}
\title{Settings for MS2 parameters.}
\usage{
defaultMs2Settings(...)
}
\arguments{
\item{\ldots}{named arguments that should be overwritten}
}
\value{
named \code{list} of settings
}
\description{
Settings for MS2 parameters.
}
\examples{
library("topdown")

# all default settings
defaultMs2Settings()

# overwrite AgcTarget
defaultMs2Settings(AgcTarget=c(1e4, 1e5))
}
