% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-functions.R
\name{expandMs2Conditions}
\alias{expandMs2Conditions}
\title{Expand MS2 Conditions}
\usage{
expandMs2Conditions(massList, ActivationType = c("CID", "HCD", "ETD",
  "UVPD"), ...)
}
\arguments{
\item{massList}{\code{matrix}, 2 columns (mass, z).}

\item{ActivationType}{\code{character}, \emph{ActivationType} for MS2 one of CID, HCD, ETD, or
UVPD.}

\item{...}{further named arguments, used to create the combination of
conditions.}
}
\value{
\code{data.frame}
}
\description{
Expand MS2 Conditions
}
\examples{
expandMs2Conditions(
     massList=cbind(mz=c(560.6, 700.5, 933.7), z=rep(1, 3)),
     ActivationType="CID",
     OrbitrapResolution="R120K",
     IsolationWindow=1,
     MaxITTimeInMS=200,
     Microscans=as.integer(40),
     AgcTarget=c(1e5, 5e5, 1e6),
     CIDCollisionEnergy=c(NA, seq(7, 35, 7))
)
}
