% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{runScanHeadsman}
\alias{runScanHeadsman}
\title{Create .csv/.txt files from raw files}
\usage{
runScanHeadsman(path = ".", executable = "ScanHeadsman.exe")
}
\arguments{
\item{path}{\code{character}, path to the directory containing the \code{.raw} files.}

\item{executable}{\code{character}, path to the \code{ScanHeadsman.exe} executable.}
}
\value{
Nothing. Used for its side effects.
}
\description{
This function calls \code{ScanHeadsman.exe} on a given directory containing
\code{.raw} files. \code{ScanHeadsman.exe} extracts the method and scan header data
into \code{.experiments.csv} and \code{.txt} files, respectively.
It works only on Windows.
}
\examples{
\dontrun{
runScanHeadsman("raw", executable="..\\\\ScanHeadsman.exe")
}
}
\references{
ScanHeadsman source code:
https://bitbucket.org/caetera/scanheadsman
}
\seealso{
\code{\link[=runXmlMethodChanger]{runXmlMethodChanger()}}
}
