% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitLogFiles-functions.R
\name{splitLogFile}
\alias{splitLogFile}
\title{Split Log files.}
\usage{
splitLogFile(file, idletime = 20L, sleep = 0.1, verbose = interactive())
}
\arguments{
\item{file}{file to watch}

\item{idletime}{time to wait for file changes}

\item{sleep}{time to sleep before next check}

\item{verbose}{verbose output?}
}
\description{
This function watches a Xtract logfile and split it into individual spectra
files.
}
