% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{topdown-package}
\alias{topdown}
\alias{topdown-package}
\title{Investigation of Fragmentation Conditions in Top-Down Proteomics}
\description{
The topdown package allows automatic and systemic investigation of
fragment conditions. It creates Thermo Orbitrap Fusion Lumos method files
to test hundreds of fragmentation conditions. Additionally it provides
functions to analyse and process the generated MS data and determine the
best conditions to maximise overall fragment coverage.
}
\details{
The usage of the topdown package is demonstrated in the following vignettes:
\itemize{
\item Generate .meth files prior data acquisition for the Thermo Orbitrap Fusion
Lumos MS devise: \code{vignette("data-generation", package="topdown")}.
\item How to analyse top-down fragmenation data: \code{vignette("analysis", package="topdown")}
}
}
\references{
\url{https://github.com/sgibb/topdown/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sgibb/topdown/}
  \item Report bugs at \url{https://github.com/sgibb/topdown/issues/}
}

}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\keyword{package}
