% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-TopDownExperiment.R
\name{TopDownExperiment}
\alias{TopDownExperiment}
\title{TopDownExperiment}
\usage{
TopDownExperiment(path, pattern = ".*", type = c("a", "b", "c", "x", "y",
  "z"), modifications = c(C = 57.02146), neutralLoss = defaultNeutralLoss(),
  tolerance = 1e-05, onDisk = FALSE, verbose = interactive(), ...)
}
\arguments{
\item{path}{\code{character}, path to top down files.}

\item{pattern}{\code{character}, pattern for file names.}

\item{type}{character, type of fragments (should be some more than in
MSnbase)}

\item{modifications}{modifications (see MSnbase::calculateFragments)}

\item{neutralLoss}{neutral loss (see MSnbase::calculateFragments)}

\item{tolerance}{double, tolerance to match peaks}

\item{onDisk}{logical, use \code{\linkS4class{MSnExp}} (\code{FALSE},
default) or \code{\linkS4class{OnDiskMSnExp}} (\code{TRUE}) for spectra
storage.}

\item{verbose}{logical, verbose output?}
}
\value{
TopDownExperiment object.
}
\description{
TopDownExperiment
}
