% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-deprecated.R
\name{topdownr-deprecated}
\alias{topdownr-deprecated}
\title{Deprecated functions in topdownr}
\description{
These functions are provided for compatibility with older versions
of \sQuote{MyPkg} only, and will be defunct at the next release.
}
\details{
The following functions are deprecated and will be made defunct; use
the replacement indicated below:
\itemize{
\item \code{defaultMs1Settings}: \code{\link[=expandMs1Conditions]{expandMs1Conditions()}} in combination with
\code{\link[=createExperimentsFragmentOptimisation]{createExperimentsFragmentOptimisation()}}
\item \code{defaultMs2Settings}: \code{\link[=expandTms2Conditions]{expandTms2Conditions()}} in combination with
\code{\link[=createExperimentsFragmentOptimisation]{createExperimentsFragmentOptimisation()}}
}
}
\keyword{deprecated}
\keyword{package}
