% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{createTngFusionMethFiles}
\alias{createTngFusionMethFiles}
\alias{runXmlMethodChanger}
\title{Create meth files from xml templates}
\usage{
createTngFusionMethFiles(template, xml = list.files(pattern = ".*\\\\.xml$"),
  executable = "XmlMethodChanger.exe", verbose = interactive())

runXmlMethodChanger(template, xml = list.files(pattern = ".*\\\\.xml$"),
  executable = "XmlMethodChanger.exe", verbose = interactive())
}
\arguments{
\item{template}{\code{character}, path to template \code{.meth} file.}

\item{xml}{\code{character}, vector of path to \code{.xml} files.}

\item{executable}{\code{character}, path to the \code{XmlMethodChanger.exe}
executable.}

\item{verbose}{\code{logical}, if \code{TRUE} a progress bar is shown.}
}
\value{
Nothing. Used for its side effects.
}
\description{
This function calls \code{XmlMethodChanger.exe} on all given xml files generated
with \code{\link[=writeMethodXmls]{writeMethodXmls()}}.
It works only on Windows.
}
\examples{
\dontrun{
runXmlMethodChanger(templateMeth="TMS2IndependentTemplate240Extended.meth",
                    modificationXml=list.files(pattern="^method.*\\\\.xml$"),
                    executable="..\\\\XmlMethodChanger.exe")
}
}
\references{
XmlMethodChanger source code:
https://github.com/thermofisherlsms/meth-modifications/
}
\seealso{
\code{\link[=writeMethodXmls]{writeMethodXmls()}}
}
