% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-AbstractTopDownSet.R
\docType{class}
\name{AbstractTopDownSet-class}
\alias{AbstractTopDownSet-class}
\alias{[,AbstractTopDownSet,ANY,ANY,ANY-method}
\alias{$,AbstractTopDownSet-method}
\alias{$<-,AbstractTopDownSet-method}
\alias{assayData,AbstractTopDownSet-method}
\alias{colData,AbstractTopDownSet-method}
\alias{colData}
\alias{colData<-,AbstractTopDownSet-method}
\alias{colData<-}
\alias{conditionData,AbstractTopDownSet-method}
\alias{conditionData<-,AbstractTopDownSet-method}
\alias{dim,AbstractTopDownSet-method}
\alias{dimnames,AbstractTopDownSet-method}
\alias{removeEmptyConditions,AbstractTopDownSet-method}
\alias{rowViews,AbstractTopDownSet-method}
\alias{rowViews}
\alias{show,AbstractTopDownSet-method}
\title{The AbstractTopDownSet class}
\usage{
\S4method{[}{AbstractTopDownSet,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

\S4method{$}{AbstractTopDownSet}(x, name)

\S4method{$}{AbstractTopDownSet}(x, name) <- value

\S4method{assayData}{AbstractTopDownSet}(object)

\S4method{colData}{AbstractTopDownSet}(object)

\S4method{colData}{AbstractTopDownSet}(object, ...) <- value

\S4method{conditionData}{AbstractTopDownSet}(object, ...)

\S4method{conditionData}{AbstractTopDownSet}(object, ...) <- value

\S4method{dim}{AbstractTopDownSet}(x)

\S4method{dimnames}{AbstractTopDownSet}(x)

\S4method{removeEmptyConditions}{AbstractTopDownSet}(object)

\S4method{rowViews}{AbstractTopDownSet}(object, ...)

\S4method{show}{AbstractTopDownSet}(object)
}
\arguments{
\item{x}{\code{AbstractTopDownSet}}

\item{i}{\code{numeric}, \code{logical} or \code{character}, subsetting on fragment/bond data,
names (\code{c("a1", "b1", "c1", "c2", "c3")}) and types (\code{c("c", "x")}) (or for
\linkS4class{NCBSet} bonds e.g. \code{bond001}) are supported.}

\item{j}{\code{numeric} or \code{logical}, subsetting based on condition data.}

\item{drop}{\code{logical}, currently ignored.}

\item{object}{\code{AbstractTopDownSet}}

\item{\ldots}{currently ignored.}
}
\description{
Abstract/VIRTUAL parent class for \linkS4class{TopDownSet} and \linkS4class{NCBSet} to
provide common interface.
}
\details{
This class just provides a common interface. It is not intended for direct
use by the user. Please see \linkS4class{TopDownSet} for an example usage of its
child class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{[}: Subset operator.

\item \code{$}: Accessor for columns in the \code{colData} slot.

The \code{$} simplifies the accession of a single column of the \code{colData}.

\item \code{$<-}: Setter for a column in the \code{colData} slot.

The \code{$<-} operator is used to add/replace a single column of the \code{colData}
\code{DataFrame}.

\item \code{assayData}: Accessor for the \code{assay} slot.

Returns a \link[Matrix:dgCMatrix-class]{Matrix::dgCMatrix} that stores the intensity/coverage
information of \linkS4class{AbstractTopDownSet} object.

\item \code{colData}: Accessor for the \code{colData} slot.

Returns a \link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame} that stores metadata for the
conditons/runs (columns) of the \linkS4class{AbstractTopDownSet} object.

\item \code{colData<-}: Setter for the \code{colData} slot.

Replaces metadata for the conditons/runs (columns) of the
\linkS4class{AbstractTopDownSet} object.

\item \code{conditionData}: Accessor for the \code{colData} slot.

An alias for \code{colData}.

\item \code{conditionData<-}: Setter for the \code{colData} slot.

An alias for \code{colData<-}.

\item \code{dim}: Accessor for dimensions.

Returns a \code{numeric} with number of fragments/bonds (rows) and
conditions/runs (columns).

\item \code{dimnames}: Accessor for dimension names.

Returns a \code{list} with names for the fragments/bonds (rows) and for the
conditions/runs (columns).

\item \code{removeEmptyConditions}: Remove empty conditions/runs.

Removes conditions/runs (columns) without any intensity/coverage
information from the \linkS4class{AbstractTopDownSet} object. It returns a modified
\linkS4class{AbstractTopDownSet} object.

\item \code{rowViews}: Accessor for the \code{rowViews} slot.

Depending on the implementation it returns an \linkS4class{FragmentViews}
object for \linkS4class{TopDownSet} objects or an \link[Biostrings:XStringViews]{Biostrings::XStringViews}
object for \linkS4class{NCBSet} objects.
}}

\section{Slots}{

\describe{
\item{\code{rowViews}}{\link[Biostrings:XStringViews-class]{Biostrings::XStringViews}, information about fragments/bonds
(name, type, sequence, mass, charge), see \link[Biostrings:XStringViews-class]{Biostrings::XStringViews} and
\linkS4class{FragmentViews} for details.}

\item{\code{colData}}{\link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame}, information about the MS2
experiments and the fragmentation conditions.}

\item{\code{assay}}{\link[Matrix:dgCMatrix-class]{Matrix::dgCMatrix}, intensity/coverage values of the
fragments/bonds. The rows correspond to the fragments/bonds and the
columns to the condition/run. It just stores values that are different from
zero.}

\item{\code{files}}{\code{character}, files that were imported.}

\item{\code{tolerance}}{\code{double}, tolerance in \emph{ppm} that were used for matching the
experimental \emph{m/z} values to the theoretical fragments.}

\item{\code{processing}}{\code{character}, log messages.}
}}

\seealso{
\itemize{
\item \linkS4class{TopDownSet} and \linkS4class{NCBSet} which both implement/use this
interface. These manual pages also provide some example code.
\item \linkS4class{FragmentViews} (and \link[Biostrings:XStringViews-class]{Biostrings::XStringViews}) for the row
view interface.
\item \link[Matrix:dgCMatrix-class]{Matrix::dgCMatrix} for technical details about the intensity/coverage storage.
}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
