% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-functions.R
\name{createExperimentsFragmentOptimisation}
\alias{createExperimentsFragmentOptimisation}
\title{Create fragment optimisation experiment}
\usage{
createExperimentsFragmentOptimisation(ms1, ..., groupBy = c("AgcTarget",
  "replication"), nMs2perMs1 = 10, scanDuration = 0,
  replications = 2, randomise = TRUE)
}
\arguments{
\item{ms1}{\code{data.frame}, MS1 settings.}

\item{...}{further named arguments with \code{data.frame}s containing the MS2
settings.}

\item{groupBy}{\code{character}, group experiments by columns in the MS2
\code{data.frame}s. The columns have to be present in all \code{data.frame}s. Each
group will be written to its own XML file.}

\item{nMs2perMs1}{\code{integer}, how many MS2 scans should be run after a MS1
scan?}

\item{scanDuration}{\code{double}, if greater than zero (e.g. \code{scanDuration=0.5})
the Start/EndTime are overwritten with a duration of \code{scanDuration}. If
\code{scanDuration} is zero (default) Start/EndTime are not overwritten.}

\item{replications}{\code{integer}, number of replications.}

\item{randomise}{\code{logical}, should the MS2 scan settings randomised?}
}
\value{
\code{list}, able to be written via \code{\link[xml2:as_xml_document]{xml2::as_xml_document()}}
}
\description{
Create fragment optimisation experiment
}
