% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-functions.R
\name{defaultMs1Settings}
\alias{defaultMs1Settings}
\alias{defaultMs2Settings}
\title{Settings for MS1/2 parameters.}
\usage{
defaultMs1Settings(...)

defaultMs2Settings(...)
}
\arguments{
\item{\ldots}{named arguments that should be overwritten.}
}
\value{
A named \code{list} of settings.
}
\description{
These functions are deprecated. Use \code{\link[=createExperimentsFragmentOptimisation]{createExperimentsFragmentOptimisation()}}
instead.
}
