% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default-functions.R
\name{defaultMs1Settings}
\alias{defaultMs1Settings}
\alias{defaultMs2Settings}
\title{Settings for MS1/2 parameters.}
\usage{
defaultMs1Settings(...)

defaultMs2Settings(...)
}
\arguments{
\item{\ldots}{named arguments that should be overwritten.}
}
\value{
A named \code{list} of settings.
}
\description{
These functions are deprecated. Use \code{\link[=createExperimentsFragmentOptimisation]{createExperimentsFragmentOptimisation()}}
instead.
}
\examples{
# all default MS1 settings
defaultMs1Settings()

# overwrite FirstMass and set AgcTarget
defaultMs1Settings(FirstMass=100, AgcTarget=c(1e4, 1e5))

# all default MS2 settings
defaultMs2Settings()

# overwrite AgcTarget
defaultMs2Settings(AgcTarget=c(1e4, 1e5))
}
