% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment-functions.R
\name{expandMs1Conditions}
\alias{expandMs1Conditions}
\title{Expand MS1 Conditions}
\usage{
expandMs1Conditions(...)
}
\arguments{
\item{...}{further named arguments, used to create the combination of
conditions.}

\item{type}{\code{character}, \emph{ActivationType} for MS2 one of CID, HCD, ETD, or
UVPD.}
}
\value{
\code{data.frame}
}
\description{
Expand MS1 Conditions
}
