% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/functions-FragmentViews.R,
%   R/methods-FragmentViews.R
\docType{class}
\name{FragmentViews-class}
\alias{FragmentViews-class}
\alias{FragmentViews}
\alias{show,FragmentViews-method}
\title{The FragmentViews class}
\usage{
FragmentViews(sequence, mass, type, z = 1L, start = NULL, end = NULL,
  width = NULL, names = NULL)

\S4method{show}{FragmentViews}(object)
}
\arguments{
\item{sequence}{\code{character}/\code{AAString}, complete protein/peptide sequence.}

\item{mass}{\code{double}, mass of the fragments, same length as
\code{start}/\code{end}/\code{width}.}

\item{type}{\code{character}, type of the fragments, same length as
\code{start}/\code{end}/width`.}

\item{z}{\code{integer}, charge of the fragments, length one or same length as
\code{start}/\code{end}/width`.}

\item{start}{\code{integer}, start positions of the fragments. At least two of
\code{start}/\code{end}/width` has to be given.}

\item{end}{\code{integer}, end positions of the fragments. At least two of
\code{start}/\code{end}/width` has to be given.}

\item{width}{\code{integer}, width positions of the fragments. At least two of
\code{start}/\code{end}/width` has to be given.}

\item{names}{\code{character}, names of the fragments, same length as
\code{start}/\code{end}/width`.}

\item{object}{FragmentViews}
}
\description{
The FragmentViews class is a basic container for storing a set of views
(start/end locations) on the same peptides/protein sequence. Additionally
it keeps information about mass, type and charge of the fragments.
}
\details{
FragmentViews extends \link[Biostrings:XStringViews-class]{Biostrings::XStringViews}.
In short it combines an \link[IRanges:IRanges-class]{IRanges::IRanges} object to store start/end
location on a sequence, an \link[Biostrings:AAString]{Biostrings::AAString} object.
}
\section{Functions}{
\itemize{
\item \code{FragmentViews}: Constructor

In general it is not necessary to call the constructor manually. See
\code{\link[=readTopDownFiles]{readTopDownFiles()}} instead.
}}

\examples{
library("topdown")

# constructor
fv <- FragmentViews("ACE", start=1, width=1:3, names=paste0("b", 1:3),
                    mass=c(72.04439, 232.07504, 361.11763),
                    type="b", z=1)
fv
}
\seealso{
\link[Biostrings:XStringViews-class]{Biostrings::XStringViews}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
