% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{NCBSet-class}
\alias{NCBSet-class}
\title{The NCBSet class}
\description{
The NCBSet class is a container for a top-down proteomics
experiment similar to the \linkS4class{TopDownSet} but instead of intensity values
it just stores the information if a bond is covered by a
N-terminal (encoded as \code{1}), a C-terminal (encoded as \code{2})
and/or both fragments (encoded as \code{3}).
}
\seealso{
\linkS4class{TopDownSet}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
