% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{runScanHeadsman}
\alias{runScanHeadsman}
\title{Create .csv/.txt files from raw files}
\usage{
runScanHeadsman(path = ".", executable = "ScanHeadsman.exe")
}
\arguments{
\item{path}{\code{character}, path to the directory containing the raw files.}

\item{executable}{\code{character}, path to the
\sQuote{ScanHeadsman.exe} executable.}
}
\value{
Nothing. Used for its side effects.
}
\description{
This function calls \sQuote{ScanHeadsman.exe} on a given directory containing
RAW files.
It works only on Windows.
}
\references{
ScanHeadsman source code:
\url{https://bitbucket.org/caetera/scanheadsman}
}
\seealso{
\code{\link{XmlMethodChanger}}
}
