% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{topdown-package}
\alias{topdown-package}
\title{Helper functions for topdown proteomics.}
\description{
Some xml helper functions for topdown proteomics.
}
\references{
\url{https://github.com/sgibb/topdown/}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
\keyword{package}
