% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/functions-TopDownSet.R
\docType{class}
\name{TopDownSet-class}
\alias{TopDownSet-class}
\alias{readTopDownFiles}
\title{The TopDownSet class}
\usage{
readTopDownFiles(path, pattern = ".*", type = c("a", "b", "c", "x", "y",
  "z"), modifications = c(C = 57.02146), adducts = data.frame(),
  neutralLoss = MSnbase::defaultNeutralLoss(), tolerance = 1e-05,
  dropNonInformativeColumns = TRUE, verbose = interactive(), ...)
}
\arguments{
\item{path}{\code{character}, path to directory that contains the top-down files.}

\item{pattern}{\code{character}, a filename pattern, the default \code{.*} means all
files.}

\item{type}{\code{character}, type of fragments, currently \emph{a-c} and \emph{x-z} are
supported, see \code{\link[MSnbase:calulateFragments]{MSnbase::calulateFragments()}} for details.}

\item{adducts}{\code{data.frame}, with 3 columns, namely: mass, name, to, see
details section.}

\item{neutralLoss}{\code{list}, neutral loss that should be applied, see
\code{\link[MSnbase:calulateFragments]{MSnbase::calulateFragments()}} and \code{\link[MSnbase:defaultNeutralLoss]{MSnbase::defaultNeutralLoss()}} for
details.}

\item{tolerance}{double, tolerance in \emph{ppm} that is used to match the
theoretical fragments with the observed ones.}

\item{dropNonInformativeColumns}{logical, should columns with just one
identical value across all runs be removed?}

\item{verbose}{logical, verbose output?}

\item{modification}{\code{double}, named vector with modifications that should be
applied. The name hast to be an amino acid that should be replaced, see
\code{\link[MSnbase:calulateFragments]{MSnbase::calulateFragments()}} for details.}
}
\value{
A \code{TopDownSet} object.
}
\description{
The TopDownSet class is a container for a whole top-down proteomics
experiment.
}
\details{
See \code{vignette("TopDown", package="topdown")} for a detailed example how to
work with \code{TopDownSet} objects.

\code{readTopDownFiles} reads and processes all top-down files, namely:
\itemize{
\item \code{.fasta} (protein sequence)
\item \code{.mzML} (spectra)
\item \code{.experiments.csv} (method/fragmentation conditions)
\item \code{.txt} (scan header information)
}

\code{adducts}: \emph{Thermo's Xtract} allows some mistakes in deisotoping, mostly it
allows \code{+/- C13-C12} and \code{+/- H+}. The \code{adducts} argument takes a
\code{data.frame} with the \code{mass} to add, the \code{name} that should assign to these
new fragments and an information \code{to} whom the modification should be
applied, e.g. for \code{H+} on \code{z}, \code{data.frame(mass=1.008, name="zpH", to="z")}.

\emph{Please note:} The \code{adducts} are added to the output of
\code{\link[MSnbase:calculateFragments]{MSnbase::calculateFragments()}}. That has some limitations, e.g.
neutral loss calculation could not be done in \link{topdown-package}. If neutral
loss should be applied on adducts you have to create additional rows, e.g.:
\code{data.frame(mass=c(1.008, 1.008), name=c("cpH", "cpH_"), to=c("c", "c_"))}.
}
\section{Functions}{
\itemize{
\item \code{readTopDownFiles}: Constructor/Read top-down files.

It creates an \linkS4class{TopDownSet} object and is its only constructor.
}}

\section{Slots}{

\describe{
\item{\code{rowViews}}{\linkS4class{FragmentViews}, information about fragments
(name, type, sequence, mass, charge), see \linkS4class{FragmentViews} for details.}

\item{\code{colData}}{\link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame}, information about the MS2
experiments and the fragmentation conditions.}

\item{\code{assay}}{\link[Matrix:dgCMatrix-class]{Matrix::dgCMatrix}, intensity values of the fragments. The
rows corresponding to the fragments and the columns to the condition/run. It
just stores values that are different from zero.}

\item{\code{files}}{\code{character}, files that were imported.}

\item{\code{tolerance}}{\code{double}, tolerance in \emph{ppm} that were used for matching the
experimental mz values to the theoretical fragments.}

\item{\code{processing}}{\code{character}, log messages.}
}}

\examples{
library("topdown")

if (require("topdowndata")) {
  # add H+ to z and no neutral loss of water
  tds <- readTopDownFiles(topdowndata::h2aPath(),
                          adducts=data.frame(mass=1.008, name="zpH", to="z"),
                          neutralLoss=MSnbase::defaultNeutralLoss(
                           disableWaterLoss=c("Cterm", "D", "E", "S", "T")),
                          tolerance=25e-6)
}
}
\seealso{
\linkS4class{FragmentViews}

\code{\link[MSnbase:calulateFragments]{MSnbase::calulateFragments()}}, \code{\link[MSnbase:defaultNeutralLoss]{MSnbase::defaultNeutralLoss()}}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
