% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{TopDownSet-class}
\alias{TopDownSet-class}
\title{The TopDownSet class}
\description{
The TopDownSet class is a container for a whole top-down proteomics
experiment.
}
\section{Slots}{

\describe{
\item{\code{rowViews}}{\linkS4class{FragmentViews}, information about fragments
(name, type, sequence, mass, charge), see \linkS4class{FragmentViews} for details.}

\item{\code{colData}}{\link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame}, information about the MS2
experiments and the fragmentation conditions.}

\item{\code{assay}}{\link[Matrix:dgCMatrix-class]{Matrix::dgCMatrix}, intensity values of the fragments. The
rows corresponding to the fragments and the columns to the condition/run. It
just stores values that are different from zero.}

\item{\code{files}}{\code{character}, files that were imported.}

\item{\code{tolerance}}{\code{double}, tolerance in \emph{ppm} that were used for matching the
experimental mz values to the theoretical fragments.}

\item{\code{processing}}{\code{character}, log messages.}
}}

\seealso{
\linkS4class{FragmentViews}
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}
