% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-xml-functions.R
\name{writeMethodXmls}
\alias{writeMethodXmls}
\title{Create Orbitrap Fusion method.xml files.}
\usage{
writeMethodXmls(ms1Settings, ms2Settings, groupBy = c("replication",
  "ETDReactionTime"), replications = 2, mz, massLabeling = TRUE,
  nMs2perMs1 = 10, duration = 0.5, randomise = TRUE,
  pattern = "method_\%s.xml", verbose = interactive())
}
\arguments{
\item{ms1Settings}{\code{list}, named list of MS1 settings.}

\item{ms2Settings}{\code{list}, named list of MS2 settings.}

\item{groupBy}{\code{character}, split files by \code{groupBy} entries.}

\item{replications}{\code{integer}, number of replications of experiments.}

\item{mz}{\code{matrix}, two columns (column 1: \emph{mass}, column 2:
\emph{z}).}

\item{massLabeling}{\code{logical}, should \emph{mz} values used for ID
labeling?}

\item{nMs2perMs1}{\code{integer}, how many MS2 scans should be run after a
MS1 scan?}

\item{duration}{\code{double}, how long should the scan be?}

\item{randomise}{\code{logical}, should the MS2 scan settings randomised?}

\item{pattern}{\code{character}, file name pattern for the method.xml files.}

\item{verbose}{\code{logical}, verbose output?}
}
\value{
An invisivble \code{list} with the MS1, MS2, runtimes and mz is
 returned.
}
\description{
This function is used to create Orbitrap Fusion method files from
all combinations of a user-given set of MS1 and MS2 settings.
}
\details{
\itemize{
 \item{ms1Settings:}{A \code{list} of MS1 settings. This have to be a named
 \code{list}. Valid MS1 settings are:
   \code{c("FirstMass", "LastMass", "Microscans", "MaxITTimeInMS",
   "AgcTarget")}}
 \item{ms2Settings:}{A \code{list} of MS2 settings. This have to be a named
 \code{list}. Valid MS2 settings are:
   \code{c("ActivationType", "IsolationWindow", "EnableMultiplexIons",
           "EnableMSXIds", "MaxNoOfMultiplexIons", "OrbitrapResolution",
           "AgcTarget", "MinAgcTarget", "MaxITTimeInMS", "Microscans",
           "ETDReactionTime", "ETDReagentTarget",
           "MaximumETDReagentInjectionTime", "UseInternalCalibratedETD",
           "ETDSupplementalActivationEnergy", "ETDSupplementalActivation")}}
 \item{groupBy:}{The \code{groupBy} parameter is used to split methods into
 different files. Valid entries are all settings that could be used in
 \code{ms2Settings} and \code{"replication"}).}
 \item{massLabeling:}{The Orbitrap Fusion devices seems not to respect the
 start and end times of the runs given in the method.xml files. That's why it
 is nearly impossible to identify the run with its conditions based on the
 timings. If \code{massLabeling} is \code{TRUE} (default) the mass values given
 in \code{mz} are rounded to the first decimal place and the second to fourth
 decimal place is used as numeric identifier.}
 \item{pattern:}{The file name pattern used to name different method files.
 It must contain a \code{"\%s"} that is replaced by the conditions defined in
 \code{groupby}. }
}
}
\examples{
\dontrun{
library("topdown")

writeMethodXmls(defaultMs1Settings(FirstMass=100),
                defaultMs2Settings(),
                groupBy=c("replication", "ETDReactionTime"),
                replications=4,
                pattern="method_firstmass_100_\%s.xml")
}

}
\author{
Sebastian Gibb \email{mail@sebasitangibb.de}, Pavel V. Shliaha
\email{pavels@bmb.sdu.dk}
}
