% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.R
\name{createTngFusionMethFiles}
\alias{createTngFusionMethFiles}
\alias{runXmlMethodChanger}
\title{Create meth files from xml templates}
\usage{
createTngFusionMethFiles(template, xml = list.files(pattern = ".*\\\\.xml$"),
  executable = "XmlMethodChanger.exe", verbose = interactive())

runXmlMethodChanger(template, xml = list.files(pattern = ".*\\\\.xml$"),
  executable = "XmlMethodChanger.exe", verbose = interactive())
}
\arguments{
\item{template}{\code{character}, path to template meth file.}

\item{xml}{\code{character}, vector of path to xml files.}

\item{executable}{\code{character}, path to the
\sQuote{XmlMethodChanger.exe} executable.}

\item{verbose}{\code{logical}, if \code{TRUE} a progress bar is shown.}
}
\value{
Nothing. Used for its side effects.
}
\description{
This function calls \sQuote{XmlMethodChanger.exe} on all given xml
files generated with \code{\link{writeMethodXmls}}.
It works only on Windows.
}
\references{
XmlMethodChanger source code:
\url{https://github.com/thermofisherlsms/meth-modifications}
}
\seealso{
\code{\link{writeMethodXmls}}
}
